Add working on chat

+++ src/main/java/com/milos/numeric/repositories/FileRepository.java
@@ -3,7 +3,7 @@ package com.milos.numeric.repositories;
 import com.milos.numeric.entities.Pdf;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepository extends JpaRepository<Pdf, Integer>
+public interface FileRepository extends JpaRepository<Pdf, Long>
 {
     public Pdf findByName(String name);
 }
