+++ src/main/java/com/milos/numeric/repositories/StudentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer>
+public interface StudentRepository extends JpaRepository<Student, Long>
 {
     public List<Student> findAllByOrderByPointsDesc();
 }
